requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
openai>=1.30.0
pillow==10.3.0
httpx==0.27.0
feedparser==6.0.11
supabase==2.4.2
pydantic==2.7.1
youtube-transcript-api==0.6.2
